requirements.txt
Flask>=2.1.0
Flask-SQLAlchemy>=3.0.0
requests>=2.28.0
beautifulsoup4>=4.11.0
praw>=7.7.0
TikTokApi>=0.10.0 
textblob>=0.7.0
selenium>=4.4.0    
openai>=0.27.0
pandas>=1.5.0
numpy>=1.23.0
nltk>=3.7
spacy>=3.4.0
wordcloud>=1.8.1
matplotlib>=3.5.0
python-dotenv>=0.21.0 
pytest>=7.1.0   
TikTokApi
playwright>=1.27.0